// types/trailblazersOfTomorrow.ts
export interface TTrailblazersOfTomorrow {
  body: {
    author: string;
    content: string;
    size: string;
  }[];
  description: string;
  subtitle: string[];
  thumbnail: string;
  videoUrl: string;
  videoUrlDekstop: string;
  videoDesktop: {
    mp4: string;
    webm: string;
  };
  title: string;
}